requirements.txt
six
werkzeug == 0.16.1
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
tabulate
coverage>=4.0.3
nose>=1.3.7
flask_testing
swagger-parser
swagger-tester